refactor: removed a not needed nested file
