+++ src/app/(home)/page.tsx
@@ -19,10 +19,6 @@ export default function Home() {
           href={RegisterRoutes.PlayerData}
           content='Cadastro do jogador'
         />
-        <ButtonNavigation
-          href={RegisterRoutes.PlayerProfile}
-          content='Perfil do jogador'
-        />
         <SignOutButton>
           <Button>Sign out</Button>
         </SignOutButton>
