+++ src/features/register-player/player-registration/FormLayout.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import { MultiStepRegisterForm } from '@/hooks/useMultiStep'
-import ButtonBack from '../../utils/button-back'
+import ButtonBack from '../utils/button-back'
 
 interface FormlayoutProps {
   children: React.ReactNode
