+++ src/features/register-player/player-registration/components/FormInputIcon.tsx
