+++ src/features/register-player/player-registration/components/FormStepOneSection.tsx
