+++ src/features/register-player/player-registration/helper/encrypton.ts
