+++ src/features/register-player/player-registration/helper/generateXYearsUntillToday.ts
