+++ src/features/register-player/player-registration/stories/FormStep1.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { FormSetpOne } from '@/features/register-player/player-registration/FormStep1/FormStepOne'
+import { FormSetpOne } from '@/features/register-player/player-registration/FormStepOne'
 
 const meta: Meta<typeof FormSetpOne> = {
   title: 'Register/FormSetpOne',
