+++ src/features/register-player/player-registration/validation.ts
